+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0 - 2021-03-10
+
+* **Breaking change:** icons passed as List
+* AnimationDirection added
+
 ## 0.4.1 - 2021-02-28
 
 * Create travis QA